[refact] 예약 단위 선택 버튼, 예약 버튼 자식 컴포넌트(StoreLayout)로 이동
왜냐면 예약 관련 버튼들이 StoreLayout 컴포넌트에서 사용하는 변수에 의존적이기 때문. 부모 컴포넌트에 있는 변수는 일절 사용하지 않음
